requirements.txt
# Telegram Bot Core
python-telegram-bot==22.5

# Google AI SDK
google-genai==1.52.0

# System Utilities for Health Check (/status command)
psutil==7.2.0
requests==2.32.5

# Configuration management
python-dotenv==1.2.1